apps/ai-agents/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
python-dotenv
openai-agents
